backend/requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
passlib[bcrypt]
pandas
numpy
rapidfuzz
joblib
python-jose[cryptography]
openpyxl
